requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
seaborn==0.12.2
apache-airflow-providers-postgres==5.5.1
